\name{AddPackage}
\alias{AddPackage}
\title{AddTwo and AddFour}
\description{Adds two or four to a number}
\usage{
  AddTwo(x)
  AddFour(x)
}
\arguments{
  \item{x} {the number to add to.}
}
\examples{
  ## add two to three
  x<-3
  y<-AddTwo(x) # x=3 => y = 5
  
  ## add four to three
  x<-3
  y<-AddFour(3) # x=3 => y = 7
  
}
